+++ client/app/components/group/group.controller.js
