+++ client/app/components/group/group.styl
