+++ client/app/components/group/groupDetail/description/description.component.js
