+++ client/app/components/group/groupDetail/description/description.controller.js
