+++ client/app/components/group/groupDetail/description/description.html
