+++ client/app/components/group/groupDetail/description/description.spec.js
