+++ client/app/components/group/groupDetail/members/members.component.js
