+++ client/app/components/group/groupDetail/members/members.controller.js
