+++ client/app/components/group/groupDetail/members/members.spec.js
