+++ client/app/components/group/groupDetail/pickups/pickups.component.js
