+++ client/app/components/group/groupDetail/pickups/pickups.controller.js
