+++ client/app/components/group/groupDetail/pickups/pickups.html
