+++ client/app/components/group/groupDetail/pickups/pickups.spec.js
