+++ client/app/components/group/groupDetail/stores/stores.component.js
