+++ client/app/components/group/groupDetail/stores/stores.controller.js
