+++ client/app/components/group/groupDetail/stores/stores.html
