+++ client/app/components/group/groupDetail/stores/stores.js
