+++ client/app/components/group/groupDetail/stores/stores.spec.js
