+++ client/app/components/group/store/_storeDetailMap/storeDetailMap.styl
