+++ client/app/components/group/store/storeDetail.component.js
