+++ client/app/components/group/store/storeDetail.controller.js
