+++ client/app/components/group/store/storeDetail.styl
